Cleanup the gstreamer providing stage-packages
The shortwave stage-packages essentially only provide gstreamer plugins
and their dependencies.

Moved these stage-packages to a separate 'gstreamer-plugins' part making
it easier to perform an additional clean up during build and prime of
everything that is already provided by core22 and gnome extension.

Limited the stage-packages to only list the gstreamer plugins packages
and leave it up to the package manager to decide on additional required packages.

Signed-off-by: Dries Deschout <dries.deschout@dodeco.eu>